+++ src/Index/FixedArrayIndex.php
@@ -1,0 +1,252 @@
+<?php
+/**
+ *
+ * MIT License
+ *
+ * Copyright (C) 2021  Kirill Yegorov https://github.com/k-samuel
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ *
+ */
+
+declare(strict_types=1);
+
+namespace KSamuel\FacetedSearch\Index;
+
+use KSamuel\FacetedSearch\Indexer\IndexerInterface;
+
+/**
+ * Simple faceted index
+ * @package KSamuel\FacetedSearch
+ */
+class FixedArrayIndex implements IndexInterface
+{
+    /**
+     * Index data
+     * @var array<int|string,array<int|string,array<int>|\SplFixedArray<int>>>
+     */
+    protected $data = [];
+    /**
+     * @var null|array<int,bool>
+     */
+    private $idMapCache = null;
+    /**
+     * @var array<IndexerInterface>
+     */
+    protected $indexers = [];
+    /**
+     * @var bool
+     */
+    protected $isCompact = false;
+
+    /**
+     * Add record to index
+     * @param int $recordId
+     * @param array<int|string,array<int,mixed>> $recordValues -  ['fieldName'=>'fieldValue','fieldName2'=>['val1','val2']]
+     * @return bool
+     */
+    public function addRecord(int $recordId, array $recordValues): bool
+    {
+        if($this->isCompact){
+            throw new \RuntimeException('Cannot add record in readonly mode. Use $index->writeMode()');
+        }
+
+        $this->resetLocalCache();
+
+        foreach ($recordValues as $fieldName => $values) {
+            if (!is_array($values)) {
+                $values = [$values];
+            }
+
+            $values = array_unique($values);
+
+            if (isset($this->indexers[$fieldName])) {
+                if (!isset($this->data[$fieldName])) {
+                    $this->data[$fieldName] = [];
+                }
+                if (!$this->indexers[$fieldName]->add($this->data[$fieldName], $recordId, $values)) {
+                    return false;
+                }
+            } else {
+                foreach ($values as $value) {
+                    if (is_bool($value)) {
+                        $value = (int)$value;
+                    }
+                    if (is_float($value)) {
+                        $value = (string)$value;
+                    }
+                    $this->data[$fieldName][$value][] = $recordId;
+                }
+            }
+        }
+        return true;
+    }
+
+    /**
+     * Get facet data.
+     * @return array<int|string,array<int|string,array<int>|\SplFixedArray<int>>>
+     */
+    public function getData(): array
+    {
+        return $this->data;
+    }
+
+    private function resetLocalCache(): void
+    {
+        $this->idMapCache = null;
+    }
+    /**
+     * Get index data. Can be used for storing it to DB
+     * @return array<int|string,array<int|string,array<int>>>
+     */
+    public function export(): array
+    {
+        if ($this->isCompact) {
+            $this->writeMode();
+        }
+        /**
+         * @var array<int|string,array<int|string,array<int>>>
+         */
+         return $this->data;
+    }
+
+    /**
+     * Enable write mode, don't forget to commit changes
+     */
+    public function writeMode(): void
+    {
+        foreach ($this->data as &$value) {
+            /**
+             * @var \SplFixedArray<int> $recordList
+             */
+            foreach ($value as &$recordList) {
+                if($recordList instanceof \SplFixedArray){
+                    $recordList = $recordList->toArray();
+                }
+            }
+        }
+        $this->isCompact = false;
+    }
+
+    public function commitChanges() : void
+    {
+        foreach ($this->data as &$value) {
+            /**
+             * @var array<int>|\SplFixedArray<int> $recordList
+             */
+            foreach ($value as &$recordList) {
+                if(is_array($recordList)){
+                    $recordList = \SplFixedArray::fromArray($recordList);
+                }
+            }
+        }
+        $this->isCompact = true;
+    }
+
+    /**
+     * Set index data. Can be used for restoring from DB
+     * @param array<int|string,array<int|string,array<int>>> $data
+     */
+    public function setData(array $data): void
+    {
+        $this->isCompact = false;
+        $this->data = $data;
+        $this->resetLocalCache();
+        $this->commitChanges();
+    }
+
+    /**
+     * Get field data section from index
+     * @param string $fieldName
+     * @return array<int|string,array<int>|\SplFixedArray<int>>
+     */
+    public function getFieldData(string $fieldName): array
+    {
+        return $this->data[$fieldName] ?? [];
+    }
+
+
+    /**
+     * Get all records from index
+     * @return array<int>
+     */
+    public function getAllRecordId(): array
+    {
+        return array_keys($this->getAllREcordIdMap());
+    }
+
+    /**
+     * Get all records from index as map [$id1=>true,...]
+     * @return array<int,bool>
+     */
+    public function getAllRecordIdMap(): array
+    {
+        if ($this->idMapCache !== null) {
+            return $this->idMapCache;
+        }
+
+        $result = [];
+        foreach ($this->data as $values) {
+            foreach ($values as $list) {
+                foreach ($list as $v) {
+                    $result[$v] = true;
+                }
+            }
+        }
+        /**
+         * @var array<int,bool> $result
+         */
+
+        $this->idMapCache = $result;
+        return $result;
+    }
+
+    /**
+     * Add specialized indexer for field
+     * @param string $fieldName
+     * @param IndexerInterface $indexer
+     */
+    public function addIndexer(string $fieldName, IndexerInterface $indexer): void
+    {
+        $this->indexers[$fieldName] = $indexer;
+    }
+
+    /**
+     * @param string $field
+     * @param mixed $value
+     * @return int
+     */
+    public function getRecordsCount(string $field, $value): int
+    {
+        if (!isset($this->data[$field][$value])) {
+            return 0;
+        }
+        return count($this->data[$field][$value]);
+    }
+
+    /**
+     * Check if field exists
+     * @param string $fieldName
+     * @return bool
+     */
+    public function hasField(string $fieldName): bool
+    {
+        return isset($this->data[$fieldName]);
+    }
+}