+++ tests/benchmark/Bench/ArrayIndex/S1mBench.php
@@ -2,14 +2,14 @@
 
 declare(strict_types=1);
 
-namespace KSamuel\FacetedSearch\Tests\Benchmark\Bench;
+namespace KSamuel\FacetedSearch\Tests\Benchmark\Bench\ArrayIndex;
 
 use KSamuel\FacetedSearch\Filter\FilterInterface;
 use KSamuel\FacetedSearch\Index;
 use KSamuel\FacetedSearch\Search;
 use KSamuel\FacetedSearch\Filter\ValueFilter;
 use KSamuel\FacetedSearch\Tests\Benchmark\DatasetFactory;
-use KSamuel\FacetedSearch\Tests\Benchmark\SearchBench;
+use KSamuel\FacetedSearch\Tests\Benchmark\ArrayIndexBench;
 use PhpBench\Benchmark\Metadata\Annotations\BeforeMethods;
 use PhpBench\Benchmark\Metadata\Annotations\Iterations;
 use PhpBench\Benchmark\Metadata\Annotations\Revs;
@@ -19,10 +19,10 @@ use PhpBench\Benchmark\Metadata\Annotations\Revs;
  * @Revs(10)
  * @BeforeMethods({"before"})
  */
-class S300KBench extends SearchBench
+class S1mBench extends ArrayIndexBench
 {
     /**
      * @var int
      */
-    protected $dataSize = 300000;
+    protected $dataSize = 1000000;
 }