+++ tests/unit/IndexTest.php
@@ -9,7 +9,7 @@ class IndexTest extends TestCase
 {
     public function testAddRecord()
     {
-        $index = new Index();
+        $index = new Index\ArrayIndex();
         $this->assertTrue($index->addRecord(112, ['vendor'=>'Tester','price' => 100]));
         $this->assertTrue($index->addRecord(113, ['vendor'=>'Tester2','price' => 101]));
         $this->assertTrue($index->addRecord(114, ['vendor'=>'Tester2','price' => 101]));
@@ -23,6 +23,25 @@ class IndexTest extends TestCase
                 101 => [113,114]
             ]
         ], $index->getData());
+        $this->assertTrue($index->addRecord(114, ['vendor'=>'Tester2','price' => 0.15]));
+    }
+
+    public function testFixedAddRecord()
+    {
+        $index = new Index\FixedArrayIndex();
+        $this->assertTrue($index->addRecord(112, ['vendor'=>'Tester','price' => 100]));
+        $this->assertTrue($index->addRecord(113, ['vendor'=>'Tester2','price' => 101]));
+        $this->assertTrue($index->addRecord(114, ['vendor'=>'Tester2','price' => 101]));
+        $this->assertEquals([
+            'vendor' => [
+                'Tester' => [112],
+                'Tester2' => [113,114]
+            ],
+            'price' => [
+                100 =>  [112],
+                101 => [113,114]
+            ]
+        ], $index->export());
         $this->assertTrue($index->addRecord(114, ['vendor'=>'Tester2','price' => 0.15]));
     }
 }